feat: add tx fee estimation and swap for tx fee payment integration
