+++ src/utils/hooks/transaction/use-transaction.ts
@@ -1,15 +1,5 @@
-import {
-  CurrencyExt,
-  CurrencyId,
-  ExtrinsicData,
-  isCurrencyEqual,
-  LiquidityPool,
-  MultiPath,
-  newCurrencyId,
-  Trade
-} from '@interlay/interbtc-api';
+import { CurrencyExt, LiquidityPool } from '@interlay/interbtc-api';
 import { MonetaryAmount } from '@interlay/monetary-js';
-import { SubmittableExtrinsic } from '@polkadot/api/types';
 import { ExtrinsicStatus } from '@polkadot/types/interfaces';
 import { ISubmittableResult } from '@polkadot/types/types';
 import { mergeProps } from '@react-aria/utils';
@@ -23,6 +13,7 @@ import { useGetLiquidityPools } from '../api/amm/use-get-liquidity-pools';
 import { getExtrinsic, getStatus } from './extrinsics';
 import { Transaction, TransactionActions, TransactionArgs } from './types';
 import { useTransactionNotifications } from './use-transaction-notifications';
+import { estimateTransactionFee, wrapWithTxFeeSwap } from './utils/fee';
 import { submitTransaction } from './utils/submit';
 
 type TransactionResult = { status: 'success' | 'error'; data: ISubmittableResult; error?: Error };
@@ -32,7 +23,6 @@ type ExecuteArgs<T extends Transaction> = {
   execute<D extends Transaction = T>(...args: TransactionArgs<D>): void;
   // Similar to execute but returns a promise which can be awaited.
   executeAsync<D extends Transaction = T>(...args: TransactionArgs<D>): Promise<TransactionResult>;
-  estimateFee<D extends Transaction = T>(...args: TransactionArgs<D>): Promise<MonetaryAmount<CurrencyExt>>;
 };
 
 type ExecuteTypeArgs<T extends Transaction> = {
@@ -48,111 +38,19 @@ type ReactQueryUseMutationResult = Omit<
   'mutate' | 'mutateAsync'
 >;
 
+type FeeResultType<T extends Transaction> = {
+  currency: CurrencyExt;
+  value: MonetaryAmount<CurrencyExt> | undefined;
+  onChangeFeeCurrency: (currency: CurrencyExt) => void;
+  estimateFee<D extends Transaction = T>(...args: TransactionArgs<D>): Promise<void>;
+};
+
 type UseTransactionResult<T extends Transaction> = {
   reject: (error?: Error) => void;
   isSigned: boolean;
+  fee: FeeResultType<T>;
 } & ReactQueryUseMutationResult &
   ExecuteFunctions<T>;
-
-const constructSwapPathPrimitive = (path: MultiPath): Array<CurrencyId> => {
-  const inputCurrency = newCurrencyId(window.bridge.api, path[0].input);
-  return [inputCurrency, ...path.map(({ output }) => newCurrencyId(window.bridge.api, output))];
-};
-
-// Recursively double input amount until the trade with higher than minimum output
-// amount is found.
-const getOptimalTradeForTxFeeSwap = (
-  minOutputAmount: MonetaryAmount<CurrencyExt>,
-  inputAmount: MonetaryAmount<CurrencyExt>,
-  pools: Array<LiquidityPool>
-): Trade => {
-  const trade = window.bridge.amm.getOptimalTrade(inputAmount, minOutputAmount.currency, pools);
-  if (trade === null) {
-    // TODO: handle - trade path not found, not possible to do the swap
-    throw new Error();
-  }
-  if (trade?.outputAmount.lt(minOutputAmount)) {
-    // If the output amount is lower than txFee double the input currency amount and check again.
-    return getOptimalTradeForTxFeeSwap(minOutputAmount, inputAmount.mul(2), pools);
-  }
-  return trade;
-};
-
-const getTxFeeSwapData = async (
-  nativeTxFee: MonetaryAmount<CurrencyExt>,
-  feeCurrency: CurrencyExt,
-  baseExtrinsic: SubmittableExtrinsic<'promise'>,
-  pools: Array<LiquidityPool>
-): Promise<{ swapPathPrimitive: Array<CurrencyId>; inputAmount: MonetaryAmount<CurrencyExt> }> => {
-  // First we construct reverse direction trade to get estimated swap path and amount
-  const reverseDirectionTrade = window.bridge.amm.getOptimalTrade(nativeTxFee, feeCurrency, pools);
-  if (reverseDirectionTrade === null) {
-    // TODO: handle - trade path not found, not possible to do the swap
-    throw new Error();
-  }
-  // Final native token transaction fee is estimated for base extrinsic wrapped in multiTransactionPayment call.
-  // NOTE: We assume here the reverse direction trade has similar weight.
-  const reverseDirectionExtrinsic = window.bridge.api.tx.multiTransactionPayment.withFeeSwapPath(
-    constructSwapPathPrimitive(reverseDirectionTrade.path),
-    reverseDirectionTrade.outputAmount.toString(true),
-    baseExtrinsic
-  );
-  const withSwapTxFee = await window.bridge.transaction.getFeeEstimate(reverseDirectionExtrinsic);
-  const { inputAmount, path } = getOptimalTradeForTxFeeSwap(
-    withSwapTxFee.mul(1.5),
-    reverseDirectionTrade.outputAmount,
-    pools
-  );
-  const swapPathPrimitive = constructSwapPathPrimitive(path);
-
-  return { inputAmount, swapPathPrimitive };
-};
-
-const getTransactionFee: (
-  feeCurrency: CurrencyExt,
-  pools: Array<LiquidityPool>
-) => (params: TransactionActions) => Promise<MonetaryAmount<CurrencyExt>> = (feeCurrency, pools) => async (params) => {
-  const baseExtrinsicData = await getExtrinsic(params);
-  const nativeTxFee = await window.bridge.transaction.getFeeEstimate(baseExtrinsicData.extrinsic);
-
-  if (isCurrencyEqual(feeCurrency, GOVERNANCE_TOKEN)) {
-    return nativeTxFee;
-  }
-
-  const { inputAmount: swapTxFee } = await getTxFeeSwapData(
-    nativeTxFee,
-    feeCurrency,
-    baseExtrinsicData.extrinsic,
-    pools
-  );
-
-  return swapTxFee;
-};
-
-const wrapWithTxFeeSwap = async (
-  feeCurrency: CurrencyExt,
-  baseExtrinsicData: ExtrinsicData,
-  pools: Array<LiquidityPool>
-): Promise<ExtrinsicData> => {
-  if (isCurrencyEqual(feeCurrency, GOVERNANCE_TOKEN)) {
-    return baseExtrinsicData;
-  }
-  const nativeTxFee = await window.bridge.transaction.getFeeEstimate(baseExtrinsicData.extrinsic);
-
-  const { swapPathPrimitive, inputAmount } = await getTxFeeSwapData(
-    nativeTxFee,
-    feeCurrency,
-    baseExtrinsicData.extrinsic,
-    pools
-  );
-  const wrappedCall = window.bridge.api.tx.multiTransactionPayment.withFeeSwapPath(
-    swapPathPrimitive,
-    inputAmount.toString(true),
-    baseExtrinsicData.extrinsic
-  );
-  console.log(wrappedCall.toHex());
-  return { extrinsic: wrappedCall };
-};
 
 const mutateTransaction: (
   feeCurrency: CurrencyExt,
@@ -174,7 +72,6 @@ type UseTransactionOptions = Omit<
   customStatus?: ExtrinsicStatus['type'];
   onSigning?: (variables: TransactionActions) => void;
   showSuccessModal?: boolean;
-  feeCurrency?: CurrencyExt;
 };
 
 // The three declared functions are use to infer types on diferent implementations
@@ -192,6 +89,8 @@ function useTransaction<T extends Transaction>(
   const { state } = useSubstrate();
 
   const [isSigned, setSigned] = useState(false);
+  const [feeCurrency, setFeeCurrency] = useState(GOVERNANCE_TOKEN);
+  const [feeEstimate, setFeeEstimate] = useState<MonetaryAmount<CurrencyExt>>();
 
   const { showSuccessModal, customStatus, ...mutateOptions } =
     (typeof typeOrOptions === 'string' ? options : typeOrOptions) || {};
@@ -217,7 +116,7 @@ function useTransaction<T extends Transaction>(
   const { data: pools } = useGetLiquidityPools();
 
   const { mutate, mutateAsync, ...transactionMutation } = useMutation(
-    mutateTransaction(options?.feeCurrency || GOVERNANCE_TOKEN, pools || []),
+    mutateTransaction(feeCurrency, pools || []),
     optionsProp
   );
 
@@ -283,13 +182,18 @@ function useTransaction<T extends Transaction>(
   };
 
   const handleEstimateFee = useCallback(
-    (...args: Parameters<UseTransactionResult<T>['estimateFee']>) => {
+    async (...args: Parameters<FeeResultType<T>['estimateFee']>) => {
       const params = getParams(args);
 
-      return getTransactionFee(options?.feeCurrency || GOVERNANCE_TOKEN, pools || [])(params);
+      const fee = await estimateTransactionFee(feeCurrency, pools || [])(params);
+      setFeeEstimate(fee);
     },
-    [options?.feeCurrency, pools, getParams]
+    [feeCurrency, pools, getParams]
   );
+
+  const handleFeeCurrencyChange = useCallback((currency: CurrencyExt) => {
+    setFeeCurrency(currency);
+  }, []);
 
   return {
     ...transactionMutation,
@@ -297,7 +201,12 @@ function useTransaction<T extends Transaction>(
     reject: handleReject,
     execute: handleExecute,
     executeAsync: handleExecuteAsync,
-    estimateFee: handleEstimateFee
+    fee: {
+      currency: feeCurrency,
+      value: feeEstimate,
+      onChangeFeeCurrency: handleFeeCurrencyChange,
+      estimateFee: handleEstimateFee
+    }
   };
 }
